+++ CHANGES.rst
@@ -8,6 +8,8 @@ Changelog of threedigrid-builder
 
 - Added command-line interface (optionally installable via [cli]).
 
+- Fix error for models without CrossSectionDefinitions.
+
 
 1.4.0 (2022-09-21)
 ------------------